add test based on simple transform example
