+++ ods_tools/odtf/examples/simple_transform/csv/config.yaml
@@ -3,7 +3,7 @@ input:
 output:
   path: ./output.csv
 mapping:
-  path: ./mapping.yml
+  path: ../mapping.yml
 batch_size: 150000
 write_header: true
 absolute_path_database: ""
