+++ ods_tools/odtf/examples/simple_transform/sqlite/simple.sql
@@ -1,0 +1,1 @@
+SELECT * FROM location
